+++ src/Patches/RoundManagerPatch.cs
@@ -7,6 +7,7 @@ namespace SCP1507.Patches;
 [HarmonyPatch(typeof(RoundManager))]
 internal class RoundManagerPatch
 {
+    
     [HarmonyPatch("UnloadSceneObjectsEarly")]
     [HarmonyPostfix]
     private static void PostFixStart(RoundManager __instance)