+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,24 @@
 ## 1.0.0
 
 - Initial release
+
+## 1.1.0
+- Added the configuration setup!
+- MAXIMUM_NUMBER_OF_FLAMINGO :  Added a config for the maximum number of flamingos
+- CAN_SPAWN_ALPHA : Whether flamingos can spawn alpha flamingos
+- CHANCE_SPAWN_NEW_FLAMING : Chances for a new flamingo to spawn
+- RARITY_ALPHA : Rarity of the alpha flamingo
+- CHANCE_SPAWN_NEW_ALPHA : Chances for a new alpha to be spawned
+- DAMAGE_ALPHA : Damage dealt by the alpha
+- DAMAGE_NORMAL : Damage dealt by normal flamingo's
+- ANGER_FOR_ANGRY_FLAMINGO  : How much anger a player must get for the horde to hate you
+- CUSTOM_HAT : Whether flamingo's have custom hat or not
+- Fixed some small Lethal Quantities bug!
+- ## 1.1.1
+- Fixed some naming error in the configurations
+- Adjusted base rarity of alpha spawn 100 --> 40 ( I didn't mean for it to be 100, did an error while taking my configs, so sorry for flamingo's army)
+
+## 1.1.2
+
+- Fixed a bunch of error for the flamingo in the logs ( Didn't affect gameplay )
+- Added chicken Hat  ( For Takey birthday)